+++ Convert/Public/ConvertFrom-ByteArrayToBase64.ps1
@@ -28,6 +28,7 @@
 function ConvertFrom-ByteArrayToBase64
 {
     [CmdletBinding(HelpUri = 'http://convert.readthedocs.io/en/latest/functions/ConvertFrom-ByteArrayToBase64/')]
+    [Alias('ConvertFrom-ByteArrayToBase64String')]
     param
     (
         [Parameter(